Add a new method that reverses the dependency order for 'add task by name' helper, and use that in the openapi project. (#14)
The open api class generator plugin in some cases (such as with new build conventions plugins) doesn't find the tasks when dependencies are requested. This commit uses the new methodology for adding those task dependencies, but requires a new method to do so.

No changes necessary to use this build.