Update contract plugins to be functional inside a service repo and update versioning logic (#31)
This allows for the inclusion of a contract module (openapi or json) inside a service project. Allows the maintenance of multiple versions of the contract as separate modules inside that service project.

Because I encountered conflicts between something in the contract process and something in our spring builds, this also updates the consolidation process to be classpath-isolated for openapi contacts. And because I encountered some conflicts in how versions were handled in service repositories, this drops the axion plugin and unifies all version handling in the different conventions plugins.

Also creates constants for https://projectronin.atlassian.net/browse/DASH-7098